Created LineChart component.
Very similar to PieChart component, left it grayscale since color seemed a bit too abundant in the UI. Again, like PieChart, do not tamper with the responsive setting. It allows the chart to auto update with changes to it's event list. No changes were needed to Calendar.css since the chart-container formatting used on the PieChart could be reused.